API and GUI programs now share models and a context class
