+++ Data/DataDbContext.cs
@@ -19,6 +19,10 @@ namespace lunch_project.Classes
         public DbSet<School> Schools { get; set; }
 
 
+        public DataDbContext(DbContextOptions<DataDbContext> options) : base(options) //do not use this to directly create an instance of this class
+        {
+        }
+
         /**<summary>Configures connection to the database, automatically called for each instance</summary>
          */
         protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder) 
@@ -31,13 +35,7 @@ namespace lunch_project.Classes
         }
 
 
-        /**<summary>Configures the model for the database</summary>
-         */
-        /*protected override void OnModelCreating(ModelBuilder modelBuilder)
-        {
-            //Fluent API stuff
-        }
-        */
+
 
     }
 }
