--- lunch_project/Classes/LunchContext.cs
@@ -1,33 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.SqlServer;
-
-
-namespace lunch_project
-{
-    /**<summary>This DbContext class allows the GUI program to query the database</summary>
-     */
-    public class LunchContext: DbContext
-    {
-        public DbSet<Student> Students { get; set; }
-        public DbSet<FoodItem> FoodItems { get; set; }
-        public DbSet<Transaction> Transactions { get; set; }
-        public DbSet<School> Schools { get; set; }
-
-        protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
-        {
-            if (!optionsBuilder.IsConfigured)
-            {
-                //optionsBuilder.UseSqlServer(connectionString)
-                //base.OnConfiguring(optionsBuilder);
-            }
-        }
-
-
-
-    }
-}
